Access ITIS indirectly via NPS web server; pull NPS taxon info; closes #8
This is a fundamental change in the way taxonomic info is retrieved. The `itistools` dependency is no longer necessary. Rather an NPS IRMA web service is accessed based on scientific name. Attempts are made to retrieve accepted taxa in the event the SN associated with an observation is no longer accepted/valid.